pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["fniad <artkamproject@gmail.com>"]
readme = "README.md"
packages = [{include = "habit_tracker"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
rest-framework-simplejwt = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
